Implemented suggested fix for navto rule
