+++ assets/src/apps/authoring/components/Modal/diagnostics/DiagnosticMessage.tsx
@@ -39,3 +39,5 @@ export const DiagnosticMessage: React.FC<Message> = (props) => {
 
   return <Message {...props} />;
 };
+
+export default DiagnosticMessage;
