+++ assets/src/apps/authoring/components/Modal/diagnostics/SolutionProps.ts
@@ -1,0 +1,5 @@
+export interface SolutionProps {
+  type?: string;
+  suggestion: string;
+  onClick: (val: string) => void;
+}
